Split upload and effective event CLI responsibilities
Amp-Thread-ID: https://ampcode.com/threads/T-01a01330-fd5f-70dd-b3b7-526874075977
