+++ internal/cli/cli.go
@@ -2,40 +2,12 @@
 package cli
 
 import (
-	"bytes"
-	"context"
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
-	"os"
-	"os/exec"
-	"os/signal"
-	"path/filepath"
-	"runtime"
-	"sort"
 	"strings"
-	"syscall"
 
-	buildkitepipeline "github.com/buildkite/buildkite-gha/internal/buildkite"
-	"github.com/buildkite/buildkite-gha/internal/compatibility"
-	"github.com/buildkite/buildkite-gha/internal/compiler"
-	"github.com/buildkite/buildkite-gha/internal/plan"
 	gharuntime "github.com/buildkite/buildkite-gha/internal/runtime"
-	"github.com/buildkite/buildkite-gha/internal/telemetry"
 	"github.com/buildkite/buildkite-gha/internal/transport"
-	"github.com/buildkite/buildkite-gha/internal/workflow"
-)
-
-const (
-	legacyRuntimeQueue            = "hosted"
-	legacyTargetQueueEnvironment  = "BUILDKITE_GHA_TARGET_QUEUE"
-	legacyRuntimeImageEnvironment = "BUILDKITE_GHA_RUNTIME_IMAGE"
-	maxWebhookMetadataBytes       = 25 << 20
-	workflowCheckSummaryLimit     = 65535
-	workflowCheckSummaryNotice    = "\n\n_Additional diagnostics omitted at the provider check summary size limit._\n"
 )
 
 // Run executes the command and returns its process exit code.
@@ -99,925 +71,4 @@ func commandVersion(clientVersion string) string {
 		return "dev"
 	}
 	return clientVersion
-}
-
-func upload(args []string, stdout, stderr io.Writer, version string, agent transport.Agent) int {
-	return uploadFromPlatform(runtime.GOOS, runtime.GOARCH, args, stdout, stderr, version, agent)
-}
-
-func uploadFromPlatform(goos, goarch string, args []string, stdout, stderr io.Writer, version string, agent transport.Agent) int {
-	platform, err := importerPlatform(goos, goarch)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: %v\n", err)
-		return 1
-	}
-	uploadArguments, err := parseUploadArgs(args)
-	if err != nil {
-		return usageError(stderr, "upload: %v", err)
-	}
-	uploadArguments.importerPlatform = platform
-	return uploadParsed(uploadArguments, stdout, stderr, version, agent)
-}
-
-func uploadParsed(uploadArguments parsedUploadArgs, stdout, stderr io.Writer, version string, agent transport.Agent) int {
-	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
-	defer stop()
-	return uploadParsedContext(ctx, uploadArguments, stdout, stderr, version, agent)
-}
-
-func uploadParsedContext(ctx context.Context, uploadArguments parsedUploadArgs, stdout, stderr io.Writer, version string, agent transport.Agent) int {
-	workflowOperands, eventPath := uploadArguments.workflowOperands, uploadArguments.eventPath
-	importerStep := os.Getenv("BUILDKITE_STEP_KEY")
-	if os.Getenv("BUILDKITE") != "true" || strings.TrimSpace(importerStep) == "" {
-		return usageError(stderr, "upload: BUILDKITE=true and BUILDKITE_STEP_KEY are required")
-	}
-	for _, retired := range []string{legacyTargetQueueEnvironment, legacyRuntimeImageEnvironment} {
-		if os.Getenv(retired) != "" {
-			return usageError(stderr, "upload: %s is no longer supported; configure runner profiles with --runner-queue and --runner-image, or with the plugin runners array", retired)
-		}
-	}
-	out := newProcessingOutput(ctx, "upload", "text", stderr, stderr, agent)
-	out.plugin = uploadArguments.pluginAcquisition != nil
-	if uploadArguments.telemetry != nil {
-		out.observe = uploadArguments.telemetry.observe
-	}
-	var eventSource []byte
-	var eventOrigin effectiveEventOrigin
-	var eventLoadErr error
-	if eventPath != "" {
-		eventSource, eventOrigin, eventLoadErr = loadEffectiveEventSource(ctx, eventPath, agent)
-		if parsedEvent, parseErr := compiler.ParseEvent(eventSource); eventLoadErr == nil && parseErr == nil {
-			out.sourceLinks = sourceLinksForEvent(parsedEvent)
-		}
-	} else if buildEvent, buildEventErr := buildkiteEventSource(os.Getenv); buildEventErr == nil {
-		if parsedEvent, parseErr := compiler.ParseEvent(buildEvent); parseErr == nil {
-			out.sourceLinks = sourceLinksForEvent(parsedEvent)
-		}
-	}
-	var workflows []workflowInput
-	var err error
-	if uploadArguments.explicitWorkflowPaths {
-		workflows, err = expandExplicitWorkflowPaths(workflowOperands)
-	} else {
-		workflows, err = resolveWorkflowOperands(workflowOperands)
-	}
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: %v\n", err)
-		return 1
-	}
-	runnableWorkflowCount := 0
-	for i := range workflows {
-		workflows[i].Source, err = os.ReadFile(workflows[i].Path)
-		if err != nil {
-			report := compatibility.EnvironmentProcessingReport(workflows[i].Path, hostedProfile, "workflow input could not be read")
-			return out.fail(report, fmt.Errorf("read workflow %s: %w", workflows[i].CanonicalPath, err))
-		}
-		parsed, parseErr := workflow.Parse(workflows[i].Path, workflows[i].Source)
-		if parseErr != nil {
-			_, _ = validatedProcessingReport(out, workflows[i].Path, hostedProfile, workflows[i].Source, nil, false)
-			return 1
-		}
-		workflows[i].ReusableOnly = parsed.ReusableOnly()
-		workflows[i].Name = parsed.Name
-		workflows[i].Triggers = parsed.Triggers
-		if !workflows[i].ReusableOnly {
-			runnableWorkflowCount++
-		}
-	}
-	if runnableWorkflowCount == 0 {
-		_, _ = fmt.Fprintln(stderr, "buildkite-gha: upload: workflow paths matched only reusable workflow_call workflows; there is nothing to upload")
-		return 1
-	}
-	for _, input := range workflows {
-		if input.ReusableOnly {
-			continue
-		}
-		validation, validationErr := compiler.Validate(input.Path, input.Source)
-		if validation.RuntimeMatrixBoundary {
-			report := compatibility.InitialProcessingReport(input.Path, hostedProfile, false, validation, validationErr)
-			report.Result = "incompatible"
-			_ = out.write(report)
-			return 1
-		}
-	}
-	if eventPath == "" {
-		eventSource, eventOrigin, eventLoadErr = loadEffectiveEventSource(ctx, eventPath, agent)
-	}
-	if eventLoadErr != nil {
-		for _, input := range workflows {
-			if !input.ReusableOnly {
-				return out.fail(compatibility.EventInputProcessingReport(input.Path, hostedProfile, input.Source, "event input could not be acquired"), eventLoadErr)
-			}
-		}
-		return 1
-	}
-	effectiveEvent, eventParseErr := newEffectiveEvent(eventSource, eventOrigin, os.Getenv)
-	if eventParseErr != nil {
-		for _, input := range workflows {
-			if !input.ReusableOnly {
-				_, _ = validatedProcessingReport(out, input.Path, hostedProfile, input.Source, eventSource, true)
-				return 1
-			}
-		}
-		return 1
-	}
-	populateChangedPaths(&effectiveEvent.TriggerSnapshot, effectiveEvent.Event, effectiveEvent.Origin, workflows)
-	processingReports := make([]compatibility.ProcessingReport, len(workflows))
-	for i := range workflows {
-		if workflows[i].ReusableOnly {
-			continue
-		}
-		selection, triggerErr := selectWorkflowTrigger(workflows[i].Triggers, effectiveEvent)
-		if triggerErr != nil {
-			workflows[i].Applicable = true
-			workflows[i].TriggerCondition = effectiveEvent.TriggerExpressions.EventPredicate
-			processingReports[i] = triggerFailureProcessingReport(workflows[i], triggerErr)
-			continue
-		}
-		if workflows[i].PathFiltersError != "" && selection.AnnotationReason == "" {
-			workflows[i].Applicable = true
-			workflows[i].TriggerCondition = effectiveEvent.TriggerExpressions.EventPredicate
-			processingReports[i] = triggerFailureProcessingReport(workflows[i], &buildkitepipeline.UnsupportedPathFiltersError{
-				Event: effectiveEvent.Event.Event, Reason: workflows[i].PathFiltersError,
-			})
-			continue
-		}
-		workflows[i].Applicable = selection.Applicable
-		workflows[i].TriggerCondition = selection.Condition
-		workflows[i].SkipReason = selection.SkipReason
-		workflows[i].AnnotationReason = selection.AnnotationReason
-	}
-	for i, input := range workflows {
-		if !input.Applicable || processingReportHasErrors(processingReports[i]) {
-			continue
-		}
-		validationOptions := hostedOptions("", uploadArguments.runnerTargets, nil)
-		validationOptions.StepKeyNamespace = input.StepKeyNamespace
-		validation, validationErr := compiler.ValidateEventWithOptions(input.Path, input.Source, effectiveEvent.Source, validationOptions)
-		processingReports[i] = compatibility.InitialProcessingReport(input.Path, hostedProfile, true, validation, validationErr)
-		if validationErr != nil {
-			processingReports[i].Result = "incompatible"
-		}
-	}
-	executablePath, executableContents, distributionDigest, err := executable()
-	if err != nil {
-		for i, input := range workflows {
-			if !input.Applicable {
-				continue
-			}
-			processingReports[i].AddEnvironmentFailure("compiler executable could not be inspected")
-			processingReports[i].Result = "indeterminate"
-			_ = out.write(processingReports[i])
-		}
-		return 1
-	}
-	importerDistribution := runtimeDistribution{contents: executableContents, digest: distributionDigest}
-	requiredPlatforms := make(map[compiler.Platform]bool, 2)
-	for i, input := range workflows {
-		if !input.Applicable || processingReportHasErrors(processingReports[i]) {
-			continue
-		}
-		platforms, platformErr := requiredRuntimePlatforms(input.Path, input.Source, effectiveEvent.Source, "", uploadArguments.runnerTargets)
-		if platformErr != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: %v\n", platformErr)
-			return 1
-		}
-		for platform := range platforms {
-			requiredPlatforms[platform] = true
-		}
-	}
-	if uploadArguments.pluginAcquisition != nil {
-		runtimeDistributions, acquireErr := uploadArguments.pluginAcquisition.acquire(ctx, requiredPlatforms, uploadArguments.importerPlatform, importerDistribution)
-		if acquireErr != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: plugin: %v\n", acquireErr)
-			return 1
-		}
-		return finishUpload(ctx, uploadArguments, stdout, stderr, version, agent, workflows, effectiveEvent, executablePath, distributionDigest, importerStep, processingReports, out, runtimeDistributions)
-	}
-	requiredDistributionPaths := make(map[compiler.Platform]string, len(requiredPlatforms))
-	for platform := range requiredPlatforms {
-		if path, ok := uploadArguments.runtimeDistributionPaths[platform]; ok {
-			requiredDistributionPaths[platform] = path
-		}
-	}
-	configuredDistributions, err := loadRuntimeDistributions(requiredDistributionPaths)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: %v\n", err)
-		return 1
-	}
-	runtimeDistributions := make(map[compiler.Platform]runtimeDistribution, len(requiredPlatforms))
-	for _, platform := range []compiler.Platform{compiler.PlatformLinuxAMD64, compiler.PlatformDarwinARM64} {
-		if !requiredPlatforms[platform] {
-			continue
-		}
-		if configured, ok := configuredDistributions[platform]; ok {
-			runtimeDistributions[platform] = configured
-			continue
-		}
-		if platform == uploadArguments.importerPlatform {
-			runtimeDistributions[platform] = importerDistribution
-			continue
-		}
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: runtime distribution for %s is required by the selected workflows\n", platform)
-		return 1
-	}
-	return finishUpload(ctx, uploadArguments, stdout, stderr, version, agent, workflows, effectiveEvent, executablePath, distributionDigest, importerStep, processingReports, out, runtimeDistributions)
-}
-
-func finishUpload(ctx context.Context, uploadArguments parsedUploadArgs, stdout, stderr io.Writer, version string, agent transport.Agent, workflows []workflowInput, effectiveEvent effectiveEventSelection, executablePath, distributionDigest, importerStep string, processingReports []compatibility.ProcessingReport, out processingOutput, runtimeDistributions map[compiler.Platform]runtimeDistribution) int {
-	runtimeDigests := make(map[compiler.Platform]string, len(runtimeDistributions))
-	for platform, runtimeDistribution := range runtimeDistributions {
-		runtimeDigests[platform] = runtimeDistribution.digest
-	}
-	authentication := importerJobActionSourceAuthentication(stderr)
-	generatedWorkflows := make([]buildkitepipeline.Workflow, 0, len(workflows))
-	skippedWorkflows := make([]skippedWorkflow, 0)
-	planArtifacts := make([]compiler.PlanArtifact, 0)
-	failureArtifacts := make([]transport.Artifact, 0)
-	jobCount := 0
-	for i, input := range workflows {
-		if input.ReusableOnly {
-			continue
-		}
-		if !input.Applicable {
-			label := input.Name
-			if label == "" {
-				label = input.CanonicalPath
-			}
-			groupKey := "gha-workflow-" + input.Identity
-			generatedWorkflows = append(generatedWorkflows, buildkitepipeline.Workflow{
-				GroupLabel: label,
-				GroupKey:   groupKey,
-				Event:      effectiveEvent.Event.Event,
-				SkipReason: input.SkipReason,
-			})
-			skippedWorkflows = append(skippedWorkflows, skippedWorkflow{label: label, key: groupKey, reason: input.AnnotationReason})
-			continue
-		}
-		if processingReportHasErrors(processingReports[i]) {
-			failed, artifacts := failedGeneratedWorkflow(input, effectiveEvent.Event.Event, processingReports[i], out.sourceLinks)
-			generatedWorkflows = append(generatedWorkflows, failed)
-			failureArtifacts = append(failureArtifacts, artifacts...)
-			continue
-		}
-		preflight, err := compileHostedNamespaced(ctx, input.Path, input.Source, effectiveEvent.Source, version, distributionDigest, importerStep, "", uploadArguments.runnerTargets, runtimeDigests, input.StepKeyNamespace, uploadArguments.oidc, authentication)
-		applyHostedPreflight(&processingReports[i], preflight)
-		if err != nil {
-			processingReports[i].Result = classifyHostedFailure(&processingReports[i], input.Path, err)
-			var failure *hostedFailure
-			if errors.As(err, &failure) && failure.Kind == hostedEvaluationFailure {
-				failed, artifacts := failedGeneratedWorkflow(input, effectiveEvent.Event.Event, processingReports[i], out.sourceLinks)
-				generatedWorkflows = append(generatedWorkflows, failed)
-				failureArtifacts = append(failureArtifacts, artifacts...)
-				continue
-			}
-			_ = out.write(processingReports[i])
-			return 1
-		}
-		bundle := preflight.Bundle
-		label := bundle.IR.Workflow.Name
-		if label == "" {
-			label = input.CanonicalPath
-		}
-		generated := bundle.GeneratedWorkflow
-		generated.GroupLabel = label
-		generated.GroupKey = "gha-workflow-" + input.Identity
-		generated.Event = effectiveEvent.Event.Event
-		generated.Condition = input.TriggerCondition
-		generatedWorkflows = append(generatedWorkflows, generated)
-		if input.AnnotationReason != "" {
-			skippedWorkflows = append(skippedWorkflows, skippedWorkflow{label: label, key: generated.GroupKey, reason: input.AnnotationReason})
-		}
-		planArtifacts = append(planArtifacts, bundle.Plans...)
-		jobCount += len(bundle.Plans)
-		processingReports[i].SetStage(string(compiler.StageAdmission), compatibility.Passed)
-		processingReports[i].Admission.Result = "admitted"
-		processingReports[i].Result = "admitted"
-		writeCompilerWarnings(stderr, "upload", input.CanonicalPath, bundle.IR.Warnings)
-		if uploadArguments.telemetry != nil {
-			uploadArguments.telemetry.addWarnings(bundle.IR.Warnings)
-			if bundleRunsUnprovenActions(bundle) {
-				uploadArguments.telemetry.addActionRuntimeUnknown()
-			}
-		}
-	}
-	aggregatePipeline, err := buildkitepipeline.Emit(buildkitepipeline.Pipeline{
-		CompilerStep:      importerStep,
-		EventProvider:     effectiveEvent.Event.Provider,
-		DisableRunnerUser: !uploadArguments.experimentalRunnerUser,
-		Workflows:         generatedWorkflows,
-	})
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: emit aggregate Buildkite pipeline: %v\n", err)
-		return 1
-	}
-	for i, input := range workflows {
-		if input.Applicable {
-			if uploadArguments.telemetry != nil {
-				uploadArguments.telemetry.addReportDiagnostics(processingReports[i])
-			}
-			if out.plugin {
-				_ = writePluginProcessing(stdout, processingReports[i])
-			} else {
-				_ = compatibility.WriteProcessing(stdout, "text", processingReports[i])
-			}
-			if !processingReportHasErrors(processingReports[i]) {
-				out.annotate(processingReports[i])
-			}
-		}
-	}
-	artifacts := make([]transport.Artifact, 0, len(runtimeDistributions)+len(planArtifacts)+len(failureArtifacts))
-	artifactPaths := make(map[string]struct{}, cap(artifacts))
-	for _, artifact := range failureArtifacts {
-		if _, exists := artifactPaths[artifact.Path]; exists {
-			continue
-		}
-		artifactPaths[artifact.Path] = struct{}{}
-		artifacts = append(artifacts, artifact)
-	}
-	for _, platform := range []compiler.Platform{compiler.PlatformLinuxAMD64, compiler.PlatformDarwinARM64} {
-		runtimeDistribution, ok := runtimeDistributions[platform]
-		if !ok {
-			continue
-		}
-		path, pathErr := buildkitepipeline.DistributionPath(runtimeDistribution.digest)
-		if pathErr != nil {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: %v\n", pathErr)
-			return 1
-		}
-		if _, exists := artifactPaths[path]; exists {
-			continue
-		}
-		artifactPaths[path] = struct{}{}
-		artifacts = append(artifacts, transport.Artifact{Path: path, Digest: runtimeDistribution.digest, Contents: runtimeDistribution.contents})
-	}
-	for _, jobPlan := range planArtifacts {
-		if _, exists := artifactPaths[jobPlan.Path]; exists {
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: duplicate aggregate artifact path %q\n", jobPlan.Path)
-			return 1
-		}
-		artifactPaths[jobPlan.Path] = struct{}{}
-		artifacts = append(artifacts, transport.Artifact{Path: jobPlan.Path, Digest: jobPlan.Digest, Contents: jobPlan.Contents})
-	}
-	if len(artifacts) == 0 {
-		if err := agent.UploadPipeline(ctx, aggregatePipeline); err != nil {
-			if uploadArguments.telemetry != nil {
-				uploadArguments.telemetry.setFailurePhase(telemetry.FailurePhasePipelineUpload)
-			}
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: upload pipeline: %v\n", err)
-			return 1
-		}
-		out.annotateSkippedWorkflows(effectiveEvent.Event.Event, skippedWorkflows)
-		_, _ = fmt.Fprintf(stdout, "Uploaded %d jobs from %d workflows using %s with importer %s.\n", jobCount, len(generatedWorkflows), executablePath, importerStep)
-		return 0
-	}
-	root, err := os.MkdirTemp("", "buildkite-gha-upload-")
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: create artifact root: %v\n", err)
-		return 1
-	}
-	defer func() { _ = os.RemoveAll(root) }()
-
-	if err := transport.UploadArtifacts(ctx, agent, root, artifacts, aggregatePipeline); err != nil {
-		if uploadArguments.telemetry != nil {
-			phase := telemetry.FailurePhaseArtifactUpload
-			if errors.Is(err, transport.ErrPipelineUpload) {
-				phase = telemetry.FailurePhasePipelineUpload
-			}
-			uploadArguments.telemetry.setFailurePhase(phase)
-		}
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: upload: %v\n", err)
-		return 1
-	}
-	out.annotateSkippedWorkflows(effectiveEvent.Event.Event, skippedWorkflows)
-	_, _ = fmt.Fprintf(stdout, "Uploaded %d jobs from %d workflows using %s with importer %s.\n", jobCount, len(generatedWorkflows), executablePath, importerStep)
-	return 0
-}
-
-func processingReportHasErrors(report compatibility.ProcessingReport) bool {
-	for _, diagnostic := range report.Diagnostics {
-		if diagnostic.Level == "error" {
-			return true
-		}
-	}
-	return false
-}
-
-func failedGeneratedWorkflow(input workflowInput, event string, report compatibility.ProcessingReport, sourceLinks sourceLinkContext) (buildkitepipeline.Workflow, []transport.Artifact) {
-	label := input.Name
-	if label == "" {
-		label = input.CanonicalPath
-	}
-	report.Diagnostics = append([]compatibility.Diagnostic(nil), report.Diagnostics...)
-	report.Finalize()
-	messages := make([]string, 0, len(report.Diagnostics))
-	for _, diagnostic := range report.Diagnostics {
-		message := diagnostic.Message
-		if diagnostic.Code != "" {
-			message = "[" + diagnostic.Code + "] " + message
-		}
-		var attribution []string
-		if diagnostic.Job != "" {
-			attribution = append(attribution, "job="+diagnostic.Job)
-		}
-		if diagnostic.Step != 0 {
-			attribution = append(attribution, fmt.Sprintf("step=%d", diagnostic.Step))
-		}
-		if len(attribution) != 0 {
-			message += " {" + strings.Join(attribution, ", ") + "}"
-		}
-		if diagnostic.Detail != "" {
-			message += "\n  detail: " + diagnostic.Detail
-		}
-		messages = append(messages, message)
-	}
-	_, annotation := processingAnnotation(report, sourceLinks)
-	_, checkSummary := processingAnnotationWithin(report, sourceLinks, workflowCheckSummaryLimit, workflowCheckSummaryNotice, false)
-	messageArtifact := generatedFailureArtifact("messages", ".txt", "\x1b[31m"+strings.Join(messages, "\n")+"\x1b[0m\n")
-	annotationArtifact := generatedFailureArtifact("annotations", ".html", annotation)
-	workflow := buildkitepipeline.Workflow{
-		GroupLabel: label,
-		GroupKey:   "gha-workflow-" + input.Identity,
-		Event:      event,
-		Failure: &buildkitepipeline.Failure{
-			AnnotationPath: annotationArtifact.Path,
-			MessagePath:    messageArtifact.Path,
-			Summary:        checkSummary,
-		},
-	}
-	return workflow, []transport.Artifact{messageArtifact, annotationArtifact}
-}
-
-func generatedFailureArtifact(kind, extension, contents string) transport.Artifact {
-	encoded := []byte(contents)
-	digest := transport.Digest(encoded)
-	path := ".buildkite-gha/failures/" + kind + "/" + strings.TrimPrefix(digest, "sha256:") + extension
-	return transport.Artifact{Path: path, Digest: digest, Contents: encoded}
-}
-
-func requiredRuntimePlatforms(workflowPath string, workflowSource, eventSource []byte, groupLabel string, configuredTargets map[string]compiler.RunnerTarget) (map[compiler.Platform]bool, error) {
-	preflight, err := compiler.CompileWithOptions(workflowPath, workflowSource, eventSource, hostedOptions(groupLabel, configuredTargets, nil))
-	if err != nil {
-		return nil, err
-	}
-	var ir compiler.IR
-	if err := json.Unmarshal(preflight, &ir); err != nil {
-		return nil, fmt.Errorf("decode compiler preflight: %w", err)
-	}
-	platforms := make(map[compiler.Platform]bool, 2)
-	for _, job := range ir.Jobs {
-		platform, err := configuredRunnerPlatform(job.RunsOn, configuredTargets)
-		if err != nil {
-			return nil, fmt.Errorf("resolve runtime platform for job %q: %w", job.LogicalJobID, err)
-		}
-		platforms[platform] = true
-	}
-	return platforms, nil
-}
-
-type effectiveEventOrigin string
-
-const (
-	effectiveEventFromPath    effectiveEventOrigin = "event-path"
-	effectiveEventFromWebhook effectiveEventOrigin = "buildkite-webhook"
-	effectiveEventFromBuild   effectiveEventOrigin = "buildkite-environment"
-)
-
-type effectiveEventSelection struct {
-	Source             []byte
-	Event              compiler.Event
-	Origin             effectiveEventOrigin
-	TriggerExpressions buildkitepipeline.TriggerConditionExpressions
-	TriggerSnapshot    buildkitepipeline.TriggerEventSnapshot
-}
-
-func loadEffectiveEventSource(ctx context.Context, eventPath string, agent transport.Agent) ([]byte, effectiveEventOrigin, error) {
-	if eventPath != "" {
-		source, err := os.ReadFile(eventPath)
-		return source, effectiveEventFromPath, err
-	}
-	webhook, metadataErr := agent.GetMetadataBounded(ctx, "buildkite:webhook", maxWebhookMetadataBytes+1)
-	switch {
-	case metadataErr == nil:
-		webhook = bytes.TrimSuffix(webhook, []byte("\n"))
-		if len(webhook) > maxWebhookMetadataBytes {
-			return nil, "", fmt.Errorf("buildkite:webhook exceeds %d bytes", maxWebhookMetadataBytes)
-		}
-		source, err := buildkiteWebhookEventSource(os.Getenv, webhook)
-		return source, effectiveEventFromWebhook, err
-	case errors.Is(metadataErr, transport.ErrMetadataUnavailable):
-		source, err := buildkiteEventSource(os.Getenv)
-		return source, effectiveEventFromBuild, err
-	default:
-		return nil, "", metadataErr
-	}
-}
-
-func newEffectiveEvent(source []byte, origin effectiveEventOrigin, getenv func(string) string) (effectiveEventSelection, error) {
-	event, err := compiler.ParseEvent(source)
-	if err != nil {
-		return effectiveEventSelection{}, err
-	}
-	expressions, snapshot := snapshotTriggerState(event)
-	effective := effectiveEventSelection{
-		Source: source, Event: event, Origin: origin,
-		TriggerExpressions: expressions,
-		TriggerSnapshot:    snapshot,
-	}
-	if origin == effectiveEventFromPath {
-		return effective, nil
-	}
-	predicate := "true"
-	if buildSource := strings.TrimSpace(getenv("BUILDKITE_SOURCE")); buildSource != "" {
-		predicate = "build.source == " + triggerConditionLiteral(buildSource)
-	}
-	effective.TriggerExpressions.EventPredicate = predicate
-	return effective, nil
-}
-
-func snapshotTriggerState(event compiler.Event) (buildkitepipeline.TriggerConditionExpressions, buildkitepipeline.TriggerEventSnapshot) {
-	expressions := buildkitepipeline.TriggerConditionExpressions{
-		EventPredicate:        "true",
-		Branch:                "null",
-		Tag:                   "null",
-		PullRequestBaseBranch: "null",
-		PullRequestAction:     "null",
-		MergeGroupBaseBranch:  "null",
-		MergeGroupAction:      "null",
-		ReleaseAction:         "null",
-	}
-	snapshot := buildkitepipeline.TriggerEventSnapshot{}
-	if branch, ok := strings.CutPrefix(event.Ref, "refs/heads/"); ok {
-		expressions.Branch = triggerConditionLiteral(branch)
-		snapshot.Branch = &branch
-	}
-	if tag, ok := strings.CutPrefix(event.Ref, "refs/tags/"); ok {
-		expressions.Tag = triggerConditionLiteral(tag)
-		snapshot.Tag = &tag
-	}
-	if action, ok := event.Payload["action"].(string); ok && strings.TrimSpace(action) != "" {
-		expressions.PullRequestAction = triggerConditionLiteral(action)
-		snapshot.PullRequestAction = &action
-		expressions.MergeGroupAction = triggerConditionLiteral(action)
-		snapshot.MergeGroupAction = &action
-		expressions.ReleaseAction = triggerConditionLiteral(action)
-		snapshot.ReleaseAction = &action
-	}
-	if pullRequest, ok := event.Payload["pull_request"].(map[string]any); ok {
-		if base, ok := pullRequest["base"].(map[string]any); ok {
-			if branch, ok := base["ref"].(string); ok && strings.TrimSpace(branch) != "" {
-				expressions.PullRequestBaseBranch = triggerConditionLiteral(branch)
-				snapshot.PullRequestBaseBranch = &branch
-			}
-		}
-	}
-	if mergeGroup, ok := event.Payload["merge_group"].(map[string]any); ok {
-		if baseRef, ok := mergeGroup["base_ref"].(string); ok {
-			if branch, ok := strings.CutPrefix(baseRef, "refs/heads/"); ok && branch != "" {
-				expressions.MergeGroupBaseBranch = triggerConditionLiteral(branch)
-				snapshot.MergeGroupBaseBranch = &branch
-			}
-		}
-	}
-	return expressions, snapshot
-}
-
-type workflowTriggerSelection struct {
-	Condition, SkipReason, AnnotationReason string
-	Applicable                              bool
-}
-
-func selectWorkflowTrigger(triggers []workflow.Trigger, event effectiveEventSelection) (workflowTriggerSelection, error) {
-	condition, applicable, err := buildkitepipeline.TranslateEventTriggerCondition(triggers, event.Event.Event, event.TriggerExpressions, event.TriggerSnapshot)
-	if err != nil {
-		return workflowTriggerSelection{}, err
-	}
-	annotationReason := buildkitepipeline.TriggerEventSkipReason(triggers, event.Event.Event)
-	if applicable {
-		annotationReason, err = buildkitepipeline.TriggerFilterMismatchReason(triggers, event.Event.Event, event.TriggerSnapshot)
-		if err != nil {
-			return workflowTriggerSelection{}, err
-		}
-	}
-	return workflowTriggerSelection{
-		Condition:        condition,
-		Applicable:       applicable,
-		SkipReason:       buildkitepipeline.TriggerEventSkipReason(triggers, event.Event.Event),
-		AnnotationReason: annotationReason,
-	}, nil
-}
-
-func pathFilterContextRequired(err error) bool {
-	var pathFilters *buildkitepipeline.UnsupportedPathFiltersError
-	return errors.As(err, &pathFilters) && (pathFilters.Event == "push" || pathFilters.Event == "pull_request") && pathFilters.Reason == ""
-}
-
-func triggerConditionLiteral(value string) string {
-	encoded, _ := json.Marshal(value)
-	return string(encoded)
-}
-
-func triggerFailureProcessingReport(input workflowInput, err error) compatibility.ProcessingReport {
-	report := triggerProcessingReport(input.Path, input.Source)
-	var pathFilters *buildkitepipeline.UnsupportedPathFiltersError
-	if errors.As(err, &pathFilters) {
-		message := fmt.Sprintf("%s trigger path filters cannot be translated safely. Remove paths and paths-ignore from this trigger, or move the filtering into a job or step.", upperFirst(pathFilters.Event))
-		if pathFilters.Reason != "" {
-			history := strings.ReplaceAll(pathFilters.Event, "_", "-")
-			message = fmt.Sprintf("%s trigger path filters could not be evaluated safely. Ensure the linked webhook and local checkout contain matching %s history, or remove the path filters.", upperFirst(pathFilters.Event), history)
-		}
-		err = &compiler.ProcessingFinding{
-			Stage: compiler.StagePipeline, Code: compiler.CodePipelineGeneration, Category: "compatibility",
-			Path: input.Path, Line: 1, Column: 1,
-			Message: message,
-			Detail:  pathFilters.Error(), Err: err,
-		}
-	}
-	report.AddFailure(input.Path, string(compiler.StagePipeline), compiler.CodePipelineGeneration, "compatibility", err)
-	report.Result = "incompatible"
-	return report
-}
-
-func triggerProcessingReport(path string, source []byte) compatibility.ProcessingReport {
-	parsed, _ := compiler.ParseWorkflow(path, source)
-	report := compatibility.NewProcessingReport(path, hostedProfile)
-	report.LogicalJobs = parsed.LogicalJobs
-	report.SetStage(string(compiler.StageWorkflowParsing), compatibility.Passed)
-	report.SetStage(string(compiler.StageEventValidation), compatibility.Passed)
-	for _, job := range parsed.ParsedJobs {
-		report.Jobs = append(report.Jobs, compatibility.JobResult{
-			ID: job.ID, Result: compatibility.NotEvaluated,
-			Location: &compatibility.SourceLocation{Path: job.Path, Line: job.Source.Start.Line, Column: job.Source.Start.Column},
-		})
-	}
-	return report
-}
-
-type workflowInput struct {
-	Path, CanonicalPath, Identity, StepKeyNamespace, Name string
-	Source                                                []byte
-	Triggers                                              []workflow.Trigger
-	TriggerCondition, SkipReason, AnnotationReason        string
-	PathFiltersError                                      string
-	ReusableOnly, Applicable                              bool
-}
-
-func resolveWorkflowOperands(operands []string) ([]workflowInput, error) {
-	if len(operands) != 1 {
-		return expandExplicitWorkflowPaths(operands)
-	}
-	path, err := filepath.Abs(operands[0])
-	if err != nil {
-		return nil, fmt.Errorf("resolve workflow path %q: %w", operands[0], err)
-	}
-	if err := requireRegularWorkflowFile(path, operands[0]); err != nil {
-		return expandExplicitWorkflowPaths(operands)
-	}
-	extension := filepath.Ext(path)
-	if extension != ".yml" && extension != ".yaml" {
-		return nil, fmt.Errorf("workflow path %q must end in .yml or .yaml", operands[0])
-	}
-	canonical := filepath.ToSlash(filepath.Clean(operands[0]))
-	if rootBytes, rootErr := exec.Command("git", "rev-parse", "--show-toplevel").Output(); rootErr == nil {
-		root := filepath.Clean(strings.TrimSpace(string(rootBytes)))
-		if relative, relativeErr := filepath.Rel(root, path); relativeErr == nil && relative != ".." && !strings.HasPrefix(relative, ".."+string(filepath.Separator)) {
-			canonical = filepath.ToSlash(filepath.Clean(relative))
-		}
-	}
-	return workflowInputs([]workflowInput{{Path: path, CanonicalPath: canonical}}, false)
-}
-
-func expandExplicitWorkflowPaths(operands []string) ([]workflowInput, error) {
-	if len(operands) == 0 {
-		return nil, fmt.Errorf("workflow path is required")
-	}
-	rootBytes, err := exec.Command("git", "rev-parse", "--show-toplevel").Output()
-	if err != nil {
-		return nil, fmt.Errorf("locate checked-out git repository: %w", err)
-	}
-	root := filepath.Clean(strings.TrimSpace(string(rootBytes)))
-	matches := make([]workflowInput, 0, len(operands))
-	for _, operand := range operands {
-		absolute, err := filepath.Abs(operand)
-		if err != nil {
-			return nil, fmt.Errorf("resolve workflow path %q: %w", operand, err)
-		}
-		relative, err := filepath.Rel(root, absolute)
-		if err != nil || relative == ".." || strings.HasPrefix(relative, ".."+string(filepath.Separator)) {
-			return nil, fmt.Errorf("workflow path %q is outside the checked-out git repository", operand)
-		}
-		canonical := filepath.ToSlash(filepath.Clean(relative))
-		if err := requireRegularWorkflowFile(absolute, operand); err != nil {
-			if strings.ContainsAny(operand, "*?[") {
-				return nil, fmt.Errorf("workflow list entries must be explicit paths; glob pattern %q is not allowed", operand)
-			}
-			return nil, err
-		}
-		extension := filepath.Ext(canonical)
-		if extension != ".yml" && extension != ".yaml" {
-			return nil, fmt.Errorf("workflow path %q must end in .yml or .yaml", operand)
-		}
-		output, err := exec.Command("git", "-C", root, "ls-files", "-z", "--", ":(top,literal)"+canonical).Output()
-		entries := bytes.Split(output, []byte{0})
-		if err != nil || len(entries) != 2 || string(entries[0]) != canonical || len(entries[1]) != 0 {
-			if strings.ContainsAny(operand, "*?[") {
-				return nil, fmt.Errorf("workflow list entries must be explicit paths; glob pattern %q is not allowed", operand)
-			}
-			return nil, fmt.Errorf("workflow path %q is not tracked by git", operand)
-		}
-		matches = append(matches, workflowInput{Path: filepath.Join(root, filepath.FromSlash(canonical)), CanonicalPath: canonical})
-	}
-	return workflowInputs(matches, true)
-}
-
-func requireRegularWorkflowFile(path, displayPath string) error {
-	info, err := os.Lstat(path)
-	if err != nil || !info.Mode().IsRegular() {
-		return fmt.Errorf("workflow path %q does not name a regular tracked file", displayPath)
-	}
-	return nil
-}
-
-func workflowInputs(matches []workflowInput, namespaceKeys bool) ([]workflowInput, error) {
-	sort.Slice(matches, func(i, j int) bool { return matches[i].CanonicalPath < matches[j].CanonicalPath })
-	out := matches[:0]
-	identities := make(map[string]string, len(matches))
-	for _, match := range matches {
-		if len(out) != 0 && out[len(out)-1].CanonicalPath == match.CanonicalPath {
-			continue
-		}
-		digest := sha256.Sum256([]byte(match.CanonicalPath))
-		match.Identity = hex.EncodeToString(digest[:8])
-		if other, exists := identities[match.Identity]; exists {
-			return nil, fmt.Errorf("workflow identity collision between %q and %q", other, match.CanonicalPath)
-		}
-		identities[match.Identity] = match.CanonicalPath
-		if namespaceKeys {
-			match.StepKeyNamespace = match.Identity
-		}
-		out = append(out, match)
-	}
-	return out, nil
-}
-
-type parsedUploadArgs struct {
-	workflowOperands         []string
-	explicitWorkflowPaths    bool
-	eventPath                string
-	runtimeDistributionPaths map[compiler.Platform]string
-	runnerTargets            map[string]compiler.RunnerTarget
-	oidc                     *plan.OIDCConfiguration
-	experimentalRunnerUser   bool
-	pluginAcquisition        *pluginRuntimeAcquisition
-	importerPlatform         compiler.Platform
-	telemetry                *commandTelemetryDetails
-}
-
-func uploadArgs(args []string) (workflowOperands []string, eventPath string, err error) {
-	parsed, err := parseUploadArgs(args)
-	return parsed.workflowOperands, parsed.eventPath, err
-}
-
-func parseUploadArgs(args []string) (parsedUploadArgs, error) {
-	workflowOperands := make([]string, 0, len(args))
-	runtimeDistributionPaths := make(map[compiler.Platform]string)
-	runnerQueues := make(map[string]string)
-	runnerImages := make(map[string]string)
-	eventPath := ""
-	eventPathSeen := false
-	runtimeQueue := ""
-	runtimeQueueSeen := false
-	deprecatedPrivateCheckoutSeen := false
-	experimentalRunnerUser := true
-	experimentalRunnerUserSeen := false
-	optionsEnded := false
-	for i := 0; i < len(args); i++ {
-		if optionsEnded {
-			workflowOperands = append(workflowOperands, args[i])
-			continue
-		}
-		if args[i] == "--" {
-			optionsEnded = true
-			continue
-		}
-		if args[i] == "--event-path" {
-			if eventPathSeen {
-				return parsedUploadArgs{}, fmt.Errorf("--event-path may only be specified once")
-			}
-			eventPathSeen = true
-			i++
-			if i == len(args) {
-				return parsedUploadArgs{}, fmt.Errorf("--event-path requires a path")
-			}
-			eventPath = args[i]
-			continue
-		}
-		if args[i] == "--private-checkout" {
-			if deprecatedPrivateCheckoutSeen {
-				return parsedUploadArgs{}, fmt.Errorf("--private-checkout may only be specified once")
-			}
-			deprecatedPrivateCheckoutSeen = true
-			continue
-		}
-		if args[i] == "--experimental-runner-user" || strings.HasPrefix(args[i], "--experimental-runner-user=") {
-			if experimentalRunnerUserSeen {
-				return parsedUploadArgs{}, fmt.Errorf("--experimental-runner-user may only be specified once")
-			}
-			experimentalRunnerUserSeen = true
-			value, configured := strings.CutPrefix(args[i], "--experimental-runner-user=")
-			if configured {
-				if value != "true" && value != "false" {
-					return parsedUploadArgs{}, fmt.Errorf("--experimental-runner-user must be true or false")
-				}
-				experimentalRunnerUser = value == "true"
-			}
-			continue
-		}
-		if args[i] == "--runtime-distribution" {
-			i++
-			if i == len(args) {
-				return parsedUploadArgs{}, fmt.Errorf("--runtime-distribution requires platform=absolute-path")
-			}
-			platformValue, path, ok := strings.Cut(args[i], "=")
-			if !ok || path == "" {
-				return parsedUploadArgs{}, fmt.Errorf("--runtime-distribution requires platform=absolute-path")
-			}
-			platform, err := compiler.ParsePlatform(platformValue)
-			if err != nil {
-				return parsedUploadArgs{}, err
-			}
-			if !filepath.IsAbs(path) {
-				return parsedUploadArgs{}, fmt.Errorf("runtime distribution for %s must use an absolute path", platform)
-			}
-			if _, exists := runtimeDistributionPaths[platform]; exists {
-				return parsedUploadArgs{}, fmt.Errorf("runtime distribution for %s may only be specified once", platform)
-			}
-			runtimeDistributionPaths[platform] = path
-			continue
-		}
-		if args[i] == "--runner-queue" || args[i] == "--runner-image" {
-			option := args[i]
-			i++
-			if i == len(args) {
-				return parsedUploadArgs{}, fmt.Errorf("%s requires runs-on=value", option)
-			}
-			label, value, ok := strings.Cut(args[i], "=")
-			if !ok || label == "" || value == "" {
-				return parsedUploadArgs{}, fmt.Errorf("%s requires runs-on=value", option)
-			}
-			canonical, _, err := supportedRunnerTarget(label)
-			if err != nil {
-				return parsedUploadArgs{}, err
-			}
-			values := runnerQueues
-			if option == "--runner-image" {
-				values = runnerImages
-			}
-			if _, duplicate := values[canonical]; duplicate {
-				return parsedUploadArgs{}, fmt.Errorf("%s for %q may only be specified once", option, canonical)
-			}
-			values[canonical] = value
-			continue
-		}
-		if args[i] == "-h" || args[i] == "--help" {
-			return parsedUploadArgs{}, fmt.Errorf("help must be requested immediately after the command")
-		}
-		if args[i] != "--runtime-queue" {
-			if strings.HasPrefix(args[i], "-") {
-				return parsedUploadArgs{}, fmt.Errorf("unknown option %q", args[i])
-			}
-			workflowOperands = append(workflowOperands, args[i])
-			continue
-		}
-		if runtimeQueueSeen {
-			return parsedUploadArgs{}, fmt.Errorf("--runtime-queue may only be specified once")
-		}
-		runtimeQueueSeen = true
-		i++
-		if i == len(args) {
-			return parsedUploadArgs{}, fmt.Errorf("--runtime-queue requires a queue")
-		}
-		runtimeQueue = args[i]
-	}
-	if len(workflowOperands) == 0 {
-		return parsedUploadArgs{}, fmt.Errorf("workflow path is required")
-	}
-	if runtimeQueueSeen && runtimeQueue != legacyRuntimeQueue {
-		return parsedUploadArgs{}, fmt.Errorf("deprecated --runtime-queue must be %q", legacyRuntimeQueue)
-	}
-	runnerTargets := make(map[string]compiler.RunnerTarget, len(runnerQueues))
-	for label, queue := range runnerQueues {
-		image := runnerImages[label]
-		canonical, target, err := configuredRunnerTarget(label, queue, image)
-		if err != nil {
-			return parsedUploadArgs{}, err
-		}
-		runnerTargets[canonical] = target
-	}
-	for label := range runnerImages {
-		if _, ok := runnerQueues[label]; !ok {
-			return parsedUploadArgs{}, fmt.Errorf("--runner-image for %q requires --runner-queue", label)
-		}
-	}
-	return parsedUploadArgs{workflowOperands: workflowOperands, eventPath: eventPath, runtimeDistributionPaths: runtimeDistributionPaths, runnerTargets: runnerTargets, experimentalRunnerUser: experimentalRunnerUser}, nil
 }
